Project Organization, duck animations, input of first take of ducks, debug for editor, restart button, gun raycats, water particles
- Edited project set up so its a little more organized. Added recources
folder and relocated prefabs into that folder. Moved model prefabs into
this folder.
- Added a duck death animation so when the duck is hit the flapping
animation is crossfaded into the death animation.
- Added first take on implementing ducks. Duck fly by adding a force to
them and they follow that route. Not to much. When duck is hit, it
falls from the sky (activates death animation), hits water or terrain,
then the duck decays into the ground for deletion.
- Changed some scripts around so that the editor can run. As well added
some comments to scripts.
- Added a restart button (debug only) so testing can be performed
quicker. Located in top left corner and is not visible like the reload
button.
- Gun instead of using bullet collision detections uses a raycast.
Raycast is more efficent and easier to use. Will be using contact potin
later so the ducks rotate when hit.
- Added a water particle effect so when bullet (raycast) hits water a
splash effect appears.
- Slowed down camera drag just a little. Use to be 0.5f no 0.55f