AI Script fixed, added first implementation of another bird
- Edited AI Manager so that ducks spwan every fifteen seconds following
ai paths. Paths are instantiated and deleted properly. So far AI works
like it is suppose to.
- AI will stop spawning when timer reaches ai spline duration (time it
takes ai to complete a path.

- Thinking about adding another bird to the mix so there is more to
shoot, but with penalties. Thinking of changing duck script to be an
abstract class that a duck script and possibly a swan (whatever the
bird may be) will inherit from.
- Need to edit ai script for this as well.