Smoothed bird model and gun model, changed minor materials, fixed animation ai bug, added deduction display
- Smoothed the bird model out so it looks a little better.
- Smoothed the gun model out so it looks a little better.
- Changed some materials, added green feathers, got rid of gun specular
material.
- Fixed Ai bug where goose would not do die animation and sometimes
ducks wont do it. (Was deleting the gameObject before it would call
animation, only would do this on last duck, explains goose problem)
- Moved HUD scor a little.
- Added a display to show user was losing points whenever they missed a
shot or shot a goose.
- Changed goos size so it is a little bigger then duck but not as big
as it was before. Instead of (2,2,2) its (1.4,1.4,1.4). Goose when
compared to gun and human arm was unrealistically big, like 2x normal
size. Made it a little smaller so it looked a little more reasonable.