Font Change, Pause Menu, Main Menu, Beginning of GameOver Sequence, High Scores
- Removed Hud system that uses GuiText for font. Now system uses 3d
Text Mesh which scales down better with different resolutions.
- Removed various #44Font colors (roughly 6 fonts) added 4 new ones
which included better graphics. Instead of using two GuiTexts (one for
front and one for backdrop, both different colors) now use one TextMesh
that uses one font style that contains backdrop within it.
- Added a Main Menu that includes its own AI, Hud, and leaderboards.
- Main menu includes two types of animations: menu animation in and
leaderboard animation in.
- Leader boards for one gameplay is added (duck hunter) which is
labeled Time Trial. Keeps track best scores for given game.
- Have not tested the build on phone yet so entering players name for
highscore may not function yet.
- Started sequence for Game over. All it does for now is store score if
it is a highscore (not displayed visually), a gameover text swoops in,
screen fades, and a quit button is available.
- Added a pause menu that freezes the game and animations. User can
toggle volume to mute on and off (NOT SAVED IF RESET GAME, need to save
in playerprefs), restart the current session, resume current session,
or quit to main menu.
- I want to change the pause menu so that it would not be the phones
back button but instead be where the clock is. This way all four
corners of the phone are occupied with a button.