Restart Bug Fix, Water Adjustments, Terrain Fix, Pausing Phone and Game Fix, Game Over Sequence
- Fixed bug that wont allow user to restart game twice. Had to deal
with event system.
- Adjusted the water textures and materials so that the water appears
to be a little darker.
- When editing terrain in Menu I forgot that it edits the terrain in
the game scene, like a prefab. Readded old terrain to game scene so now
there are two terrain prefabs, one for menu and one for game.
- Solved pausing issue. User will not be able to pause the game during
countdown nor gameover sequence. When user leaves game, pressing the
home screen button, game pauses. This only happens if the following
conditions are met, game is not already paused, gameplay is playing,
countdown is not happening, gameover is not occuring.
- Added a little more to gameover sequence. Notifies user if they made
the leaderboard. User can type their name in. Only four initials right
now. If the user enters 4 or less they are good and game over sequence
continues. If user types 5+ the game will prompt them to retry.

** Have not tested what game will do during phone call, will game pause?
** Have not tested if user presses home key when prompted to enter name
for leaderboard.