Created Art for GameOver sequence, started new GameOver sequence
- Added art files for bird logos as well as GameOver artwork.
- Added a brown #44font, used in gameover sequence.
- Started implementation of new Game Over sequence.
- Added animations for Hud to fade out.

# Game Over, so far
- Displays Game Over text
- When GameOver fades out, player rotates 180, animation fades out all
buttons and current hud on screen (combo, score, time, ammo, tigger &
zoom), backdrop fades in.
- After backdrop fades in, a statistic sheet animates in.
- Stats displays ducks hit, geese hit, best streak, accuracy, score,
and continue button.

#### Game Over, end goal to achieve
-When displaying ducks hit it takes 2 seconds, slowly increments to the
final number. After the final number is displayed, if its greater then
0 a duck sound is played.
- Duplicate same process duck went through with goose. This happens
right after duck sound goes off. (Make sure duck sound is less then 2
seconds lond so there is no overlaping.
- After goose process slowly increment streak score up.
- At same time steak score is incrementing up accurarcy number changes.
The only thing is that accuracy does not increment up, it displays a
random number 0 - 100 every 0.1 second. This process takes 3 seconds,
and finishes with final accuracy. Should take one extra second then
streak since streak is 2 and accuracy is 3.
- At same time accuracy is incrementing score is incrementing. score
increments up in 3 seconds.
- When score reaches final mark a bonus score is incremented right
below accuracy. This bonus score is the accuracy x 10. (Max is 10,000
Min is 0 points). This bonus score follows the same animation like
deduction points except its the opposite. The bonus points go into the
score and the score changes. This will be the players final score that
gors into the leaderboards.
- Once this score is calculated a continue button fades in so the user
can hit continue. Whole process of this stat menu should take roughly 8
seconds. (may change if its too quick).
- When user hits continue, everything on stat menu disappears except
for score text and user score. These text animate to top of stat menu.
- Next appears texts like bullet points. The bullet points list: "Made
Leaderboards" & "HighScore". Next to each text will be a "-".
- From here the game will wait  2 seconds to display if the user made
the leaderboards. If they did a green checkmark will replace the "-".
If not a red x will replace it. A sound of good will play if the user
made it, a bad sound will play if the user did not make it.
- Then 2 seconds later the same will happen for high score. If the user
is the top scorer a little firework will display and a "tada" sound
will play" if the user did not make a highscore a "awww" sound will
play.
- If the user did not make the leaderboard a scribble animation will go
over the highscore text and the highscore determination process will be
skipped.
- If the user made the leaderboard, after the high score sound is
played, good or bad, user will be prompt to enter their initials.
- Once user returns from initials or did not make leaderboard, the game
will delay 1 second then a restart button and quit button will appear
at the bottom of the screen.